[STRATCONN-1432] Refactoring perform method for GA4
