+++ packages/destination-actions/src/destinations/google-analytics-4/purchase/index.ts
@@ -1,8 +1,7 @@
-import { ActionDefinition, IntegrationError } from '@segment/actions-core'
+import { ActionDefinition } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
-import { verifyCurrency, convertTimestamp } from '../ga4-functions'
-import { ProductItem } from '../ga4-types'
+import { verifyCurrency, convertTimestamp, formatItems } from '../ga4-functions'
 import {
   coupon,
   currency,
@@ -51,26 +50,6 @@ const action: ActionDefinition<Settings, Payload> = {
   perform: (request, { payload }) => {
     verifyCurrency(payload.currency)
 
-    let googleItems: ProductItem[] = []
-
-    if (payload.items) {
-      googleItems = payload.items.map((product) => {
-        if (product.item_name === undefined && product.item_id === undefined) {
-          throw new IntegrationError(
-            'One of product name or product id is required for product or impression data.',
-            'Misconfigured required field',
-            400
-          )
-        }
-
-        if (product.currency) {
-          verifyCurrency(product.currency)
-        }
-
-        return product as ProductItem
-      })
-    }
-
     return request('https://www.google-analytics.com/mp/collect', {
       method: 'POST',
       json: {
@@ -84,7 +63,7 @@ const action: ActionDefinition<Settings, Payload> = {
               affiliation: payload.affiliation,
               coupon: payload.coupon,
               currency: payload.currency,
-              items: googleItems,
+              items: formatItems(payload.items),
               transaction_id: payload.transaction_id,
               shipping: payload.shipping,
               value: payload.value,
