+++ packages/destination-actions/src/destinations/google-analytics-4/viewItemList/index.ts
@@ -1,5 +1,5 @@
-import { ActionDefinition, IntegrationError } from '@segment/actions-core'
-import { verifyCurrency, convertTimestamp } from '../ga4-functions'
+import { ActionDefinition } from '@segment/actions-core'
+import { convertTimestamp, formatItems } from '../ga4-functions'
 import {
   formatUserProperties,
   user_properties,
@@ -10,7 +10,6 @@ import {
   engagement_time_msec,
   timestamp_micros
 } from '../ga4-properties'
-import { ProductItem } from '../ga4-types'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
 
@@ -51,26 +50,6 @@ const action: ActionDefinition<Settings, Payload> = {
     params: params
   },
   perform: (request, { payload }) => {
-    let googleItems: ProductItem[] = []
-
-    if (payload.items) {
-      googleItems = payload.items.map((product) => {
-        if (product.item_name === undefined && product.item_id === undefined) {
-          throw new IntegrationError(
-            'One of product name or product id is required for product or impression data.',
-            'Misconfigured required field',
-            400
-          )
-        }
-
-        if (product.currency) {
-          verifyCurrency(product.currency)
-        }
-
-        return product as ProductItem
-      })
-    }
-
     return request('https://www.google-analytics.com/mp/collect', {
       method: 'POST',
       json: {
@@ -83,7 +62,7 @@ const action: ActionDefinition<Settings, Payload> = {
             params: {
               item_list_id: payload.item_list_id,
               item_list_name: payload.item_list_name,
-              items: googleItems,
+              items: formatItems(payload.items),
               engagement_time_msec: payload.engagement_time_msec,
               ...payload.params
             }
